feat(integrations): Connect buttons on detail page
fixes #795
