+++ src/app/integrations/list/list.component.html
@@ -1,7 +1,6 @@
 <div class="integrations list">
 
   <!-- Modal Container -->
-
   <div bsModal
        #childModal="bs-modal"
        class="modal fade"
